+++ CHANGELOG.rst
@@ -16,7 +16,7 @@ Changelog
 2.3.0 (2017-12-19)
 -------------------
 * Fixed bootstrap3 templates missing custom classes
-* Added support for custom storage per form
+* Added support for custom action per form
 
 2.2.9 (2017-10-09)
 ------------------